requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
asynctest==0.13.0
pymongo==3.12.3
pyOpenSSL==21.0.0
pytest==6.2.5
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-rerunfailures==10.2
python-dateutil==2.8.0
python-utils==3.1.0
websockets==10.1
XlsxWriter==3.0.2
xlwt==1.3.0
jira==3.1.1
bunch==1.0.1
APScheduler==3.9.1
XTestRunner==1.2.0
PyYAML~=6.0
openpyxl~=3.1.1